+++ ui/src/main.js
@@ -76,6 +76,7 @@ import "./collections/oauth2/oidcOptions";
 import "./collections/oauth2/appleOptions";
 import "./logs/logsSettingsModal";
 import "./logs/logPreviewModal";
+import "./audits/auditPreviewModal";
 import { appHeader } from "./base/appHeader";
 import { initRouter } from "./router";